Refactor raycast-instant-app-switcher-extension
- Remove ESLint configuration file
- Update Prettier configuration to remove trailing commas
- Update package dependencies for ESLint, TypeScript, and Prettier
- Refactor TypeScript configuration to target ES2023 and include necessary files
- Simplify application selection logic by utilizing hooks and components for better organization and maintainability
