+++ extensions/raycast-instant-app-switcher-extension/eslint.config.js
@@ -1,0 +1,6 @@
+const { defineConfig } = require("eslint/config");
+const raycastConfig = require("@raycast/eslint-config");
+
+module.exports = defineConfig([
+  ...raycastConfig,
+]);
