+++ extensions/raycast-instant-app-switcher-extension/tsconfig.json
@@ -1,22 +1,16 @@
 {
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "include": ["src/**/*", "raycast-env.d.ts"],
   "compilerOptions": {
-    "target": "ES2020",
-    "lib": ["ES2020"],
+    "lib": ["ES2023"],
     "module": "commonjs",
-    "moduleResolution": "node",
+    "target": "ES2023",
     "strict": true,
-    "declaration": true,
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "removeComments": true,
+    "isolatedModules": true,
     "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
     "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
     "jsx": "react-jsx",
-    "jsxImportSource": "react"
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
+    "resolveJsonModule": true
+  }
 }
